+++ CHANGELOG.rst
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog <http://keepachangelog.com/>`_.
 
 ====================
+2.6.1 - 2019-10-15
+====================
+
+Added
+-----
+* Support for the Digital Assistant service
+* Support for work requests on Instance Pool operations in the Compute service
+
+====================
 2.6.0 - 2019-10-08
 ====================